backend/pyproject.toml
[tool.poetry]
name = "taxja-backend"
version = "0.1.0"
description = "Taxja - Austrian Tax Management System Backend"
authors = ["Taxja GmbH"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
pydantic = {extras = ["email"], version = "^2.5.3"}
pydantic-settings = "^2.1.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
celery = "^5.3.6"
pytesseract = "^0.3.10"
opencv-python = "^4.9.0"
pillow = "^10.2.0"
scikit-learn = "^1.4.0"
lxml = "^5.1.0"
reportlab = "^4.0.9"
boto3 = "^1.34.34"
pyotp = "^2.9.0"
qrcode = "^7.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
black = "^24.1.1"
ruff = "^0.1.14"
mypy = "^1.8.0"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
